requirements copy.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
pydantic
